Add button debounce and state changed variable
Add button state changed variable. Allows for waking of display and only triggering action if the display is active and then button is pressed.
Human readable uptime in http interface. :-)
Add code for TLi4970 current sensor to pump controller. Not currently usable due to limited pins on an ESP8266 and fixed pins for SPI clock.